+++ src/core/connected_player/functions/player_selector_to_entities.rs
@@ -7,7 +7,7 @@ use crate::core::{
 
 pub fn player_selector_to_entities(
     command_executor_entity: Entity,
-    command_executor_handle_option: Option<u32>,
+    command_executor_handle_option: Option<u64>,
     mut player_selector: &str,
     used_names: &mut ResMut<UsedNames>,
     net_console_commands: &mut EventWriter<NetConsoleCommands>,
