+++ src/core/console_commands/functions/rcon_status.rs
@@ -9,7 +9,7 @@ use crate::core::{
 
 pub fn rcon_status(
     connected_players: &mut Query<&mut ConnectedPlayer>,
-    client_handle: u32,
+    client_handle: u64,
     client_entity: Entity,
     net_console_commands: &mut EventWriter<NetConsoleCommands>,
 ) {
