+++ src/entities/counter_windows/events.rs
@@ -13,33 +13,33 @@ pub struct CounterWindowSensorCollision {
 }
 
 pub struct InputCounterWindowToggleOpen {
-    pub handle_option: Option<u32>,
+    pub handle_option: Option<u64>,
 
     pub opener: Entity,
     pub opened: u64,
 }
 pub struct CounterWindowLockOpen {
-    pub handle_option: Option<u32>,
+    pub handle_option: Option<u64>,
 
     pub locked: Entity,
     pub locker: Entity,
 }
 
 pub struct CounterWindowLockClosed {
-    pub handle_option: Option<u32>,
+    pub handle_option: Option<u64>,
 
     pub locked: Entity,
     pub locker: Entity,
 }
 
 pub struct CounterWindowUnlock {
-    pub handle_option: Option<u32>,
+    pub handle_option: Option<u64>,
 
     pub locked: Entity,
     pub locker: Entity,
 }
 
 pub struct NetCounterWindow {
-    pub handle: u32,
+    pub handle: u64,
     pub message: ReliableServerMessage,
 }
